Make compatible with new structure
