+++ src/util/background/fetchCached.js
@@ -1,7 +1,7 @@
 const DEFAULT_CACHE_TIME = 3600;
 const ONE_DAY_IN_SECONDS = 86400;
 const ONE_WEEK_IN_SECONDS = ONE_DAY_IN_SECONDS * 7;
-const MATCH_DATE_XPATH_EXPRESSION = "//div[@class='match_head']/div[@class='right']/text()";
+const MATCH_DATE_XPATH_EXPRESSION = "/html/body/div[1]/main/section[1]/div/div[2]/div[1]/span/text()";
 
 export async function fetchCached(url, cacheValidCondition = defaultCacheValidCondition, extractor = response => response.text(), header = null) {
     return new Promise((resolve, reject) => {
@@ -26,44 +26,6 @@ export async function fetchCached(url, cacheValidCondition = defaultCacheValidCo
     });
 }
 
-function getMonthNumber(monthName) {
-    switch (monthName) {
-        case "Jan":
-            return "01";
-        case "Feb":
-            return "02";
-        case "Mar":
-            return "03";
-        case "Apr":
-            return "04";
-        case "Mai":
-            return "05";
-        case "Jun":
-            return "06";
-        case "Jul":
-            return "07";
-        case "Aug":
-            return "08";
-        case "Sep":
-            return "09";
-        case "Okt":
-            return "10";
-        case "Nov":
-            return "11";
-        case "Dec":
-            return "12";
-        case "Dez":
-            return "12";
-        case "May":
-            return "05";
-        case "Oct":
-            return "10";
-        case "Mär":
-            return "03";
-    }
-}
-
-
 export function faceitExtractor(faceitResponse) {
     if (faceitResponse.ok) {
         return faceitResponse.json();
@@ -72,15 +34,11 @@ export function faceitExtractor(faceitResponse) {
     }
 }
 
-function cacheOnlyPastMatches(cacheResponse) {
+export function cacheOnlyPastMatches(cacheResponse) {
     const matchDoc = new DOMParser().parseFromString(cacheResponse.response, "text/html");
-    const matchTimeEntries = matchDoc.evaluate(MATCH_DATE_XPATH_EXPRESSION, matchDoc, null, XPathResult.UNORDERED_NODE_SNAPSHOT_TYPE, null);
-    const matchTimeAsString = matchTimeEntries.snapshotItem(0).wholeText.trim();
-    const parsedDate = matchTimeAsString.match("([0-9]{2}) ([a-zA-ZäöüßÄÖÜ]{3}) ([0-9]{4})");
-    const day = parsedDate[1];
-    const month = getMonthNumber(parsedDate[2]);
-    const year = parsedDate[3];
-    const matchDate = new Date(year, month, day);
+    const matchTimeAsString = matchDoc.evaluate(MATCH_DATE_XPATH_EXPRESSION, matchDoc, null, XPathResult.STRING_TYPE, null).stringValue;
+    const dateComponents = matchTimeAsString.split(".").map(parseInt);
+    const matchDate = new Date(dateComponents[2], dateComponents[1], dateComponents[0]);
     const twoWeeksAgo = new Date(Date.now());
     twoWeeksAgo.setDate(twoWeeksAgo.getDate()-14);
